Adding constants and other features
Number constants (or number literals) are now supported in the chunks. It's able to support up to 2^24 constants in a single chunk. I've also added support for line numbering and further optomized it to perform better than the source material implementation. It now more closely resembles a Run-length Encoding scheme rather than a continuing list of integers.